% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{tbart-package}
\alias{tbart-package}
\title{Teitz and Bart's \eqn{p}-median problem with Spatial* and Spatial*DataFrame objects
\tabular{ll}{
 Package: \tab tbart\cr
 Type: \tab Package\cr
 Version: \tab 0.9\cr
 Date: \tab 2013-06-13\cr
 License: \tab GPL (>= 2)\cr
 Maintainer: \tab Chris Brunsdon \url{mailto:brunsdon@liverpool.ac.uk} \cr
}}
\description{
Solves Teitz and Bart's \eqn{p}-median problem - given a set of
points attempts to find subset of size p such that summed distances of any
point in the set to the nearest point in p is minimised.  Although
generally effective, this algorithm does not guarantee that a globally
optimal subset is found.
}
\author{
Chris Brunsdon
}
\references{
Teitz, M. B., and P. Bart (1968), Heuristic methods for estimating generalized vertex median of a weighted graph, Operations Research, 16, 955-961.
}

